requirements.txt
Flask==3.0.0
boto3==1.28.0
python-dotenv==1.0.0
flask-cors==4.0.0

# Optional for development
pytest==7.4.0
requests==2.31.0
